refactor: pre stub eoa design
